requirements.txt
fastapi
openai
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
